feat: add the login flow
Add the login flow to the indieauth server.

After the user initialises the server they are redirected to the login
page where they can enter their profile ID and password to log in. When
the server successfully authenticates the user it will, for now, sent a
confirmation page back to the user as well as a cookie containing a JWT
that will keep the user logged in for a short period of time.