feat: add the authorization endpoint
Add the authorization endpoint for the indieauth clients to send their
authorization requests to. Beacon will then fetch the client's metadata
information using the client_id and use it to validate the authorization
request. At this point the authorization interface has not been
implemented yet so a checkpoint will be reached if the validation is
successful.