fix: created clientAuthRequest if cookie missing
The issue was that when the user is already authorized into Beacon and
a client has sent an authorization request, the server will fail the
authorization flow because the cookie for the authorization request is
never created (it is only created when the user is redirected to the
login screen).

The fix is to create the clientAuthRequest value from the URL query
values if the cookie does not exist. Otherwise the clientAuthRequest is
retrieved from the cookie as normal.