fix: reimplement the setup process
The profile information is now stored in the 'profiles' bucket in
preparation for the possible creation of more buckets (e.g. for tokens).

The setup process has been reimplemented where there is now a Setup
function that creates the 'profiles' bucket and writes the first profile
in that bucket.

The use of the 'initialized' key has been removed and the database is
determined to be initialized is when the 'profiles' bucket is present
and it has at least one profile in it.

In the server package the Server now records whether or not the database
is initialized before running. If the database is not initialized the
entrypoint middleware automatically redirects the user to the setup form
in order to complete the setup process.

PR: https://codeflow.dananglin.me.uk/apollo/beacon/pulls/17