fix: fix client auth request persistence issues
This commit fixes the issues around persistence with the client
authorization request throughout the client <-> server authorization
dance; particularly when the user is already authenticated with Beacon.

Previously a new request was only saved to a cookie if the user has
not signed in and is redirected to the login page. This is now fixed
by saving new requests to the server's cache (for 10 minutes)
immediately after it's created. The request's state key is Base64 URL
encoded and used as the key to access the request by the server.