fix: configurable JWT cookie name
Allow instance owners to configure the name of the cookie that stores
the JWT token. Previously this was set to beacon_is_great. If the name
of the cookie is not configured then the default name is used
(beacon_is_great).

Beacon supports cookie names that include any alphanumeric character
plus and of +, -, ., or _. Beacon will fail to start if the cookie
name is unsupported.

PR: https://codeflow.dananglin.me.uk/apollo/beacon/pulls/20