fix: canonicalize the client ID
Some clients do not canonicalize their ID when initiating the
authorization request (more specifically the clients do not include the
path component). Previously Beacon would automatically reject the
authorization request if the ID is not fully canonicalized, but section
3.4 of the IndieAuth specification states that 'if a URL with no path
component is ever encountered, it MUST be treated as if it had the path
/'. Therefore this change makes Beacon canonicalize the client ID
before validating it.

PR: https://codeflow.dananglin.me.uk/apollo/beacon/pulls/31